Split up DCEL related files to be more manageable
