+++ geom/dcel_test.go
